// models/IExifFile.ts
export interface IExifFile {
    altitude: string;
    deviceId: string;
    latitude: number;
    longitude: number;
    path: string;
    timestamp: string;
    lastModified: number;
    lastModifiedDate: string;
    name: string;
    size: number;
    type: string;
    webkitRelativePath: string;
  }
  